fix: se Adaptan los nombres de acuerdo al proyecto
