+++ .env
@@ -4,7 +4,7 @@ PORT = 3000
 DB_HOST = localhost
 DB_USER = root
 DB_PASSWORD =
-DB_NAME = chatbot-uvm
+DB_NAME = coindraw
 DB_PORT = 3306
