+++ config.js
@@ -6,7 +6,7 @@ export const PORT = process.env.PORT || 3000
 export const DB_HOST =  process.env.DB_HOST || 'localhost'
 export const DB_USER = process.env.DB_USER ||  'root'
 export const DB_PASSWORD = process.env.DB_PASSWORD || ''
-export const DB_NAME = process.env.DB_NAME || 'messages'
+export const DB_NAME = process.env.DB_NAME || 'coindraw'
 export const DB_PORT = process.env.DB_PORT || 3306
